requirements.txt
pyperclip
google-auth
google-auth-oauthlib
requests
